Move test RoundTrippers from jujutest -> testing
Remove unnecessary dependencies.
