+++ apiserver/provisioner/imagemetadata_test.go
@@ -11,12 +11,12 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/apiserver/provisioner"
 	"github.com/juju/juju/environs/imagemetadata"
-	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/state/cloudimagemetadata"
+	"github.com/juju/juju/testing"
 )
 
 var (
-	testRoundTripper = &jujutest.ProxyRoundTripper{}
+	testRoundTripper = &testing.ProxyRoundTripper{}
 )
 
 func init() {
@@ -27,7 +27,7 @@ func init() {
 // useTestImageData causes the given content to be served when published metadata is requested.
 func useTestImageData(files map[string]string) {
 	if files != nil {
-		testRoundTripper.Sub = jujutest.NewCannedRoundTripper(files, nil)
+		testRoundTripper.Sub = testing.NewCannedRoundTripper(files, nil)
 		imagemetadata.DefaultBaseURL = "test:"
 	} else {
 		testRoundTripper.Sub = nil
