+++ environs/imagemetadata/simplestreams_test.go
@@ -9,7 +9,7 @@ import (
 	"net/http"
 	"reflect"
 	"strings"
-	"testing"
+	stdtesting "testing"
 
 	jc "github.com/juju/testing/checkers"
 	"github.com/juju/utils"
@@ -17,9 +17,9 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/environs/imagemetadata"
-	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/environs/simplestreams"
 	sstesting "github.com/juju/juju/environs/simplestreams/testing"
+	"github.com/juju/juju/testing"
 )
 
 var live = flag.Bool("live", false, "Include live simplestreams tests")
@@ -44,7 +44,7 @@ var liveUrls = map[string]liveTestData{
 	},
 }
 
-func Test(t *testing.T) {
+func Test(t *stdtesting.T) {
 	if *live {
 		if *vendor == "" {
 			t.Fatal("missing vendor")
@@ -334,10 +334,10 @@ type signedSuite struct {
 	origKey string
 }
 
-var testRoundTripper *jujutest.ProxyRoundTripper
+var testRoundTripper *testing.ProxyRoundTripper
 
 func init() {
-	testRoundTripper = &jujutest.ProxyRoundTripper{}
+	testRoundTripper = &testing.ProxyRoundTripper{}
 	testRoundTripper.RegisterForScheme("signedtest")
 }
 
@@ -366,7 +366,7 @@ func (s *signedSuite) SetUpSuite(c *gc.C) {
 		r, sstesting.SignedMetadataPrivateKey, sstesting.PrivateKeyPassphrase)
 	c.Assert(err, jc.ErrorIsNil)
 	imageData["/signed/streams/v1/image_metadata.sjson"] = string(signedData)
-	testRoundTripper.Sub = jujutest.NewCannedRoundTripper(
+	testRoundTripper.Sub = testing.NewCannedRoundTripper(
 		imageData, map[string]int{"signedtest://unauth": http.StatusUnauthorized})
 	s.origKey = imagemetadata.SetSigningPublicKey(sstesting.SignedMetadataPublicKey)
 }
