+++ environs/tools/simplestreams_test.go
@@ -25,7 +25,6 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/environs/filestorage"
-	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/environs/simplestreams"
 	sstesting "github.com/juju/juju/environs/simplestreams/testing"
 	"github.com/juju/juju/environs/storage"
@@ -1022,10 +1021,10 @@ type signedSuite struct {
 	origKey string
 }
 
-var testRoundTripper *jujutest.ProxyRoundTripper
+var testRoundTripper *coretesting.ProxyRoundTripper
 
 func init() {
-	testRoundTripper = &jujutest.ProxyRoundTripper{}
+	testRoundTripper = &coretesting.ProxyRoundTripper{}
 	testRoundTripper.RegisterForScheme("signedtest")
 }
 
@@ -1054,7 +1053,7 @@ func (s *signedSuite) SetUpSuite(c *gc.C) {
 		r, sstesting.SignedMetadataPrivateKey, sstesting.PrivateKeyPassphrase)
 	c.Assert(err, jc.ErrorIsNil)
 	imageData["/signed/streams/v1/tools_metadata.sjson"] = string(signedData)
-	testRoundTripper.Sub = jujutest.NewCannedRoundTripper(
+	testRoundTripper.Sub = coretesting.NewCannedRoundTripper(
 		imageData, map[string]int{"signedtest://unauth": http.StatusUnauthorized})
 	s.origKey = tools.SetSigningPublicKey(sstesting.SignedMetadataPublicKey)
 }
