+++ provider/azure/internal/legacy/azure/azure_test.go
@@ -15,10 +15,9 @@ import (
 	"launchpad.net/gwacl"
 
 	"github.com/juju/juju/environs/imagemetadata"
-	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/environs/simplestreams"
 	envtesting "github.com/juju/juju/environs/testing"
-	"github.com/juju/juju/testing"
+	testing "github.com/juju/juju/testing"
 )
 
 func TestAzureProvider(t *stdtesting.T) {
@@ -33,7 +32,7 @@ type providerSuite struct {
 
 var _ = gc.Suite(&providerSuite{})
 
-var testRoundTripper = &jujutest.ProxyRoundTripper{}
+var testRoundTripper = &testing.ProxyRoundTripper{}
 
 func init() {
 	// Prepare mock http transport for overriding metadata and images output in tests.
@@ -96,5 +95,5 @@ func (s *providerSuite) makeTestMetadata(c *gc.C, ser, location string, im []*im
 		"/streams/v1/index.json":                string(index),
 		"/" + imagemetadata.ProductMetadataPath: string(products),
 	}
-	s.PatchValue(&testRoundTripper.Sub, jujutest.NewCannedRoundTripper(files, nil))
+	s.PatchValue(&testRoundTripper.Sub, testing.NewCannedRoundTripper(files, nil))
 }
