+++ provider/ec2/export_test.go
@@ -12,10 +12,10 @@ import (
 
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/imagemetadata"
-	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/environs/storage"
 	"github.com/juju/juju/instance"
 	jujustorage "github.com/juju/juju/storage"
+	"github.com/juju/juju/testing"
 )
 
 func EBSProvider() jujustorage.Provider {
@@ -71,7 +71,7 @@ func DeleteBucket(s storage.Storage) error {
 	return deleteBucket(s.(*ec2storage))
 }
 
-var testRoundTripper = &jujutest.ProxyRoundTripper{}
+var testRoundTripper = &testing.ProxyRoundTripper{}
 
 func init() {
 	// Prepare mock http transport for overriding metadata and images output in tests.
@@ -86,7 +86,7 @@ var origImagesUrl = imagemetadata.DefaultBaseURL
 // when the ec2 client asks for image data.
 func UseTestImageData(files map[string]string) {
 	if files != nil {
-		testRoundTripper.Sub = jujutest.NewCannedRoundTripper(files, nil)
+		testRoundTripper.Sub = testing.NewCannedRoundTripper(files, nil)
 		imagemetadata.DefaultBaseURL = "test:"
 		signedImageDataOnly = false
 	} else {
