+++ testing/roundtripper.go
@@ -1,7 +1,7 @@
 // Copyright 2013 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package jujutest
+package testing
 
 import (
 	"fmt"
